Allow customization of panel prefabs
This commit will allow the PanelManager singleton to be created manually by devs in a scene.

In addition this moves all the Instantiate&Resources.Load combos into the panel manager, where the "templates" can be customized by creating variants of the base prefabs and assigning them inside the PanelManager inspector.

Finally, and as an added benefit, this will cause DynamicPanels to play nicer with addressables, as the templates can be configured to _not_ use Resources.Load.